my-tkinter-app-1/requirements.txt
tkinter
requests
jsonschema
python-dotenv